+++ lib/Server.js
@@ -1,10 +1,9 @@
 "use strict";
 
 const os = require("os");
-const net = require("net");
 const path = require("path");
-const fs = require("fs");
 const url = require("url");
+const fs = require("graceful-fs");
 const ipaddr = require("ipaddr.js");
 const internalIp = require("internal-ip");
 const killable = require("killable");
@@ -1199,7 +1198,6 @@ class Server {
 
   runBonjour() {
     const bonjour = require("bonjour")();
-    const os = require("os");
 
     bonjour.publish({
       name: `Webpack Dev Server ${os.hostname()}:${this.options.port}`,
@@ -1425,6 +1423,7 @@ class Server {
     const resolveFreePortOrIPC = () => {
       if (this.options.ipc) {
         return new Promise((resolve, reject) => {
+          const net = require("net");
           const socket = new net.Socket();
 
           socket.on("error", (error) => {
